PycharmProjects/ResumeUpload/requirements.txt
streamlit
python-docx
pymupdf
openai
